Update Wanted.cpp Fix #36.
